WithXRay/requirements.txt
# -*- coding: utf-8 -*-
grpcio
grpcio-tools
protobuf
requests
